+++ packages/stg-definitions-v2/src/constant.ts
@@ -392,6 +392,12 @@ export const ASSETS: Record<TokenName, AssetConfig> = {
                 type: StargateType.Pool,
                 address: '0xdEAddEaDdeadDEadDEADDEAddEADDEAddead1111',
             },
+            [EndpointId.MONAD_V2_TESTNET]: {
+                symbol: 'WETH',
+                name: 'WETH',
+                type: StargateType.Pool,
+                address: '0xB5a30b0FDc5EA94A52fDc42e3E9760Cb8449Fb37',
+            },
             [EndpointId.OPTSEP_V2_TESTNET]: {
                 type: StargateType.Native,
             },
@@ -578,6 +584,10 @@ export const ASSETS: Record<TokenName, AssetConfig> = {
             },
             [EndpointId.MANTLESEP_V2_TESTNET]: {
                 address: '0xcC4Ac915857532ADa58D69493554C6d869932Fe6',
+                type: StargateType.Pool,
+            },
+            [EndpointId.MONAD_V2_TESTNET]: {
+                address: '0x88b8E2161DEDC77EF4ab7585569D2415a1C1055D',
                 type: StargateType.Pool,
             },
             [EndpointId.OPTSEP_V2_TESTNET]: {
@@ -812,6 +822,10 @@ export const ASSETS: Record<TokenName, AssetConfig> = {
                 address: '0xAcab8129E2cE587fD203FD770ec9ECAFA2C88080',
                 type: StargateType.Pool,
             },
+            [EndpointId.MONAD_V2_TESTNET]: {
+                address: '0xf817257fed379853cDe0fa4F97AB987181B1E5Ea',
+                type: StargateType.Pool,
+            },
             [EndpointId.OPTSEP_V2_TESTNET]: {
                 address: '0x488327236B65C61A6c083e8d811a4E0D3d1D4268',
                 type: StargateType.Pool,
@@ -954,6 +968,7 @@ export const OFT_WRAPPER: OftWrapperConfig = {
         [EndpointId.KLAYTN_V2_TESTNET]: {},
         [EndpointId.ODYSSEY_V2_TESTNET]: {},
         [EndpointId.MANTLESEP_V2_TESTNET]: {},
+        [EndpointId.MONAD_V2_TESTNET]: {},
         [EndpointId.OPTSEP_V2_TESTNET]: {},
         [EndpointId.SEPOLIA_V2_TESTNET]: {},
 
@@ -980,6 +995,7 @@ export const REWARDS: RewardsConfig = {
             [EndpointId.BSC_V2_TESTNET]: {},
             [EndpointId.OPTSEP_V2_TESTNET]: {},
             [EndpointId.MANTLESEP_V2_TESTNET]: {},
+            [EndpointId.MONAD_V2_TESTNET]: {},
             [EndpointId.SEPOLIA_V2_TESTNET]: {},
 
             //
@@ -2473,6 +2489,10 @@ export const NETWORKS: NetworksConfig = {
         tokenMessaging: DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
     },
     [EndpointId.MANTLESEP_V2_TESTNET]: {
+        creditMessaging: DEFAULT_CREDIT_MESSAGING_NETWORK_CONFIG,
+        tokenMessaging: DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
+    },
+    [EndpointId.MONAD_V2_TESTNET]: {
         creditMessaging: DEFAULT_CREDIT_MESSAGING_NETWORK_CONFIG,
         tokenMessaging: DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
     },