+++ packages/stg-evm-v2/devtools/config/testnet/asset.usdt.config.ts
@@ -24,6 +24,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const bl3Point = getAssetPoint(EndpointId.BL3_V2_TESTNET)
     const odysseyPoint = getAssetPoint(EndpointId.ODYSSEY_V2_TESTNET)
     const mantlePoint = getAssetPoint(EndpointId.MANTLESEP_V2_TESTNET)
+    const monadPoint = getAssetPoint(EndpointId.MONAD_V2_TESTNET)
 
     // And all their nodes
     const avalancheContract = await getAssetNode(avalanchePoint)
@@ -35,6 +36,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const bl3Contract = await getAssetNode(bl3Point)
     const odysseyContract = await getAssetNode(odysseyPoint)
     const mantleContract = await getAssetNode(mantlePoint)
+    const monadContract = await getAssetNode(monadPoint)
 
     // And all their connections
     const bscToEth = getAssetEdge(bscPoint, ethPoint)
@@ -45,6 +47,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const bscToBL3 = getAssetEdge(bscPoint, bl3Point)
     const bscToOdyssey = getAssetEdge(bscPoint, odysseyPoint)
     const bscToMantle = getAssetEdge(bscPoint, mantlePoint)
+    const bscToMonad = getAssetEdge(bscPoint, monadPoint)
 
     const ethToBsc = getAssetEdge(ethPoint, bscPoint)
     const ethToOpt = getAssetEdge(ethPoint, optPoint)
@@ -54,6 +57,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const ethToBL3 = getAssetEdge(ethPoint, bl3Point)
     const ethToOdyssey = getAssetEdge(ethPoint, odysseyPoint)
     const ethToMantle = getAssetEdge(ethPoint, mantlePoint)
+    const ethToMonad = getAssetEdge(ethPoint, monadPoint)
 
     const optToBsc = getAssetEdge(optPoint, bscPoint)
     const optToEth = getAssetEdge(optPoint, ethPoint)
@@ -63,6 +67,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const optToBL3 = getAssetEdge(optPoint, bl3Point)
     const optToOdyssey = getAssetEdge(optPoint, odysseyPoint)
     const optToMantle = getAssetEdge(optPoint, mantlePoint)
+    const optToMonad = getAssetEdge(optPoint, monadPoint)
 
     const arbToBsc = getAssetEdge(arbPoint, bscPoint)
     const arbToEth = getAssetEdge(arbPoint, ethPoint)
@@ -72,6 +77,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const arbToBL3 = getAssetEdge(arbPoint, bl3Point)
     const arbToOdyssey = getAssetEdge(arbPoint, odysseyPoint)
     const arbToMantle = getAssetEdge(arbPoint, mantlePoint)
+    const arbToMonad = getAssetEdge(arbPoint, monadPoint)
 
     const klaytnToBsc = getAssetEdge(klaytnPoint, bscPoint)
     const klaytnToEth = getAssetEdge(klaytnPoint, ethPoint)
@@ -81,6 +87,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const klaytnToBL3 = getAssetEdge(klaytnPoint, bl3Point)
     const klaytnToOdyssey = getAssetEdge(klaytnPoint, odysseyPoint)
     const klaytnToMantle = getAssetEdge(klaytnPoint, mantlePoint)
+    const klaytnToMonad = getAssetEdge(klaytnPoint, monadPoint)
 
     const bl3ToBsc = getAssetEdge(bl3Point, bscPoint)
     const bl3ToEth = getAssetEdge(bl3Point, ethPoint)
@@ -90,6 +97,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const bl3ToKlaytn = getAssetEdge(bl3Point, klaytnPoint)
     const bl3ToOdyssey = getAssetEdge(bl3Point, odysseyPoint)
     const bl3ToMantle = getAssetEdge(bl3Point, mantlePoint)
+    const bl3ToMonad = getAssetEdge(bl3Point, monadPoint)
 
     const odysseyToBsc = getAssetEdge(odysseyPoint, bscPoint)
     const odysseyToEth = getAssetEdge(odysseyPoint, ethPoint)
@@ -99,6 +107,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const odysseyToKlaytn = getAssetEdge(odysseyPoint, klaytnPoint)
     const odysseyToBL3 = getAssetEdge(odysseyPoint, bl3Point)
     const odysseyToMantle = getAssetEdge(odysseyPoint, mantlePoint)
+    const odysseyToMonad = getAssetEdge(odysseyPoint, monadPoint)
 
     const mantleToBsc = getAssetEdge(mantlePoint, bscPoint)
     const mantleToEth = getAssetEdge(mantlePoint, ethPoint)
@@ -108,6 +117,16 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const mantleToKlaytn = getAssetEdge(mantlePoint, klaytnPoint)
     const mantleToBL3 = getAssetEdge(mantlePoint, bl3Point)
     const mantleToOdyssey = getAssetEdge(mantlePoint, odysseyPoint)
+    const mantleToMonad = getAssetEdge(mantlePoint, monadPoint)
+
+    const monadToBsc = getAssetEdge(monadPoint, bscPoint)
+    const monadToEth = getAssetEdge(monadPoint, ethPoint)
+    const monadToOpt = getAssetEdge(monadPoint, optPoint)
+    const monadToArb = getAssetEdge(monadPoint, arbPoint)
+    const monadToKlaytn = getAssetEdge(monadPoint, klaytnPoint)
+    const monadToBL3 = getAssetEdge(monadPoint, bl3Point)
+    const monadToOdyssey = getAssetEdge(monadPoint, odysseyPoint)
+    const monadToMantle = getAssetEdge(monadPoint, mantlePoint)
 
     const avalancheToBsc = getAssetEdge(avalanchePoint, bscPoint)
     const avalancheToEth = getAssetEdge(avalanchePoint, ethPoint)
@@ -129,6 +148,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             bl3Contract,
             odysseyContract,
             mantleContract,
+            monadContract,
         ],
         connections: [
             //
@@ -142,6 +162,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             bscToBL3,
             bscToOdyssey,
             bscToMantle,
+            bscToMonad,
 
             //
             // Connections originating from ETH
@@ -154,6 +175,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             ethToBL3,
             ethToOdyssey,
             ethToMantle,
+            ethToMonad,
 
             //
             // Connections originating from OPT
@@ -166,6 +188,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             optToBL3,
             optToOdyssey,
             optToMantle,
+            optToMonad,
 
             //
             // Connections originating from ARB
@@ -178,6 +201,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             arbToBL3,
             arbToOdyssey,
             arbToMantle,
+            arbToMonad,
 
             //
             // Connections originating from AVALANCHE
@@ -202,6 +226,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             klaytnToBL3,
             klaytnToOdyssey,
             klaytnToMantle,
+            klaytnToMonad,
 
             //
             // Connections originating from BL3
@@ -214,6 +239,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             bl3ToKlaytn,
             bl3ToOdyssey,
             bl3ToMantle,
+            bl3ToMonad,
 
             //
             // Connections originating from ODYSSEY
@@ -226,6 +252,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             odysseyToKlaytn,
             odysseyToBL3,
             odysseyToMantle,
+            odysseyToMonad,
 
             //
             // Connections originating from MANTLE
@@ -238,6 +265,19 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             mantleToKlaytn,
             mantleToBL3,
             mantleToOdyssey,
+            mantleToMonad,
+
+            //
+            // Connections originating from MONAD
+            //
+            monadToBsc,
+            monadToEth,
+            monadToOpt,
+            monadToArb,
+            monadToKlaytn,
+            monadToBL3,
+            monadToOdyssey,
+            monadToMantle,
         ],
     }
 }