+++ packages/stg-evm-v2/devtools/config/testnet/rewarder.config.ts
@@ -6,7 +6,7 @@ import { EndpointId } from '@layerzerolabs/lz-definitions'
 
 import { createGetLPTokenAddresses, createGetRewardTokenAddresses } from '../../../ts-src/utils/util'
 
-import { onArb, onAvalanche, onBsc, onEth, onMantle, onOpt } from './utils'
+import { onArb, onAvalanche, onBsc, onEth, onMantle, onMonad, onOpt } from './utils'
 
 const contract = { contractName: 'StargateMultiRewarder' }
 
@@ -22,6 +22,7 @@ export default async (): Promise<OmniGraphHardhat<RewarderNodeConfig, unknown>> 
     const arbRewardTokenAddresses = await getRewardTokenAddresses(EndpointId.ARBSEP_V2_TESTNET, rewardTokens)
     const avalancheRewardTokenAddresses = await getRewardTokenAddresses(EndpointId.AVALANCHE_V2_TESTNET, rewardTokens)
     const mantleRewardTokenAddresses = await getRewardTokenAddresses(EndpointId.MANTLESEP_V2_TESTNET, rewardTokens)
+    const monadRewardTokenAddresses = await getRewardTokenAddresses(EndpointId.MONAD_V2_TESTNET, rewardTokens)
 
     const allAssets = [TokenName.USDT, TokenName.USDC, TokenName.ETH] as const
     const getLPTokenAddresses = createGetLPTokenAddresses(getEnvironment)
@@ -31,6 +32,7 @@ export default async (): Promise<OmniGraphHardhat<RewarderNodeConfig, unknown>> 
     const arbLPTokenAddresses = await getLPTokenAddresses(EndpointId.ARBSEP_V2_TESTNET, allAssets)
     const avalancheLPTokenAddresses = await getLPTokenAddresses(EndpointId.AVALANCHE_V2_TESTNET, allAssets)
     const mantleLPTokenAddresses = await getLPTokenAddresses(EndpointId.MANTLESEP_V2_TESTNET, allAssets)
+    const monadLPTokenAddresses = await getLPTokenAddresses(EndpointId.MONAD_V2_TESTNET, allAssets)
 
     // 1e18 reward per second
     const DEFAULT_REWARDS_CONFIG = {
@@ -105,6 +107,18 @@ export default async (): Promise<OmniGraphHardhat<RewarderNodeConfig, unknown>> 
                             [mantleLPTokenAddresses.USDT]: 1,
                             [mantleLPTokenAddresses.USDC]: 1,
                             [mantleLPTokenAddresses.ETH]: 1,
+                        },
+                    },
+                },
+            },
+            {
+                contract: onMonad(contract),
+                config: {
+                    allocations: {
+                        [monadRewardTokenAddresses.MOCK_A]: {
+                            [monadLPTokenAddresses.USDT]: 1,
+                            [monadLPTokenAddresses.USDC]: 1,
+                            [monadLPTokenAddresses.ETH]: 1,
                         },
                     },
                 },